Toil and trouble

+++ app/src/main/java/cs301/birthdaycake/CakeModel.java
@@ -7,7 +7,7 @@ public class CakeModel {
     public boolean ifCakeCandles = true;
     public float[] balloonCoordinates = new float[]{-1f,-1f};
 
-    float xpos;
+    double xpos;
     float ypos;
     /**
      *Constructor
